Move mutations and mappers to corresponding packages
