+++ app/src/main/AndroidManifest.xml
@@ -20,8 +20,7 @@
             android:exported="true"
             android:label="@string/app_name"
             android:screenOrientation="userPortrait"
-            android:theme="@style/Theme.BirthdayApp"
-            >
+            android:theme="@style/Theme.BirthdayApp">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
