+++ app/src/main/java/com/simply/birthdayapp/data/mappers/GetCodeResponseEntityMaper.kt
@@ -1,4 +1,4 @@
-package com.simply.birthdayapp.data.mapper
+package com.simply.birthdayapp.data.mappers
 
 import com.simply.birthdayapp.ResetPasswordEmailMutation
 import com.simply.birthdayapp.presentation.ui.models.GetCodeResponseModel
