+++ app/src/main/java/com/simply/birthdayapp/data/mappers/LoginInputMaper.kt
@@ -1,4 +1,4 @@
-package com.simply.birthdayapp.data.mapper
+package com.simply.birthdayapp.data.mappers
 
 import com.simply.birthdayapp.data.entity.LoginInputEntity
 import com.simply.birthdayapp.type.LoginInput
