+++ app/src/main/java/com/simply/birthdayapp/data/mappers/SignUpInputMaper.kt
@@ -1,4 +1,4 @@
-package com.simply.birthdayapp.data.mapper
+package com.simply.birthdayapp.data.mappers
 
 import com.simply.birthdayapp.data.entity.RegisterInputEntity
 import com.simply.birthdayapp.type.SignUpInput
