+++ app/src/main/java/com/simply/birthdayapp/data/mappers/UserEntityMapper.kt
@@ -1,4 +1,4 @@
-package com.simply.birthdayapp.data.mapper
+package com.simply.birthdayapp.data.mappers
 
 import com.simply.birthdayapp.SignUpMutation
 import com.simply.birthdayapp.presentation.ui.models.RegisteredUser
