+++ app/src/main/res/values/strings.xml
@@ -47,10 +47,8 @@
     <string name="network_error">Sorry, an error occurred registering your account, please try again.</string>
     <string name="reser_password_network_error">Sorry, an error occurred reseting your password, please try again.</string>
     <string name="sign_in_network_error">Sorry, an error occurred sign in your account, please try again.</string>
-
     <string name="get_the_code">Get The Code</string>
     <string name="set_new_password">Set New Password</string>
-    <string name="done">Done</string>
     <string name="login_success">Login Success</string>
     <string name="get_code_success">We have sent an email with a verification code, if there is an account with a provided email address</string>
 </resources>
